requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1; python_version<"3.9" and python_version>="3.8"
coverage==6.5.0
dj-database-url==1.0.0
Django==4.1.2
flake8==5.0.4
gunicorn==20.1.0
iniconfig==1.1.1
mccabe==0.7.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==4.0.0
pytest-django==4.5.2
python-dotenv==0.21.0
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
whitenoise==6.2.0
django-heroku==0.3.1
psycopg2==2.9.4
beautifulsoup4==4.11.1
django-bootstrap4==22.2
soupsieve==2.3.2.post1
django-filter==22.1
